requirements.txt
asgiref==3.8.1
black==24.8.0
certifi==2024.8.30
cfgv==3.4.0
click==8.1.7
distlib==0.3.8
dj-database-url==2.2.0
Django==5.1.1
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
filelock==3.16.0
gunicorn==23.0.0
identify==2.6.0
iniconfig==2.0.0
Markdown==3.7
mypy-extensions==1.0.0
mysqlclient==2.2.4
nodeenv==1.9.1
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.2
pluggy==1.5.0
pre-commit==3.8.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
python-dotenv==1.0.1
PyYAML==6.0.2
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
virtualenv==20.26.4
whitenoise==6.7.0
